+++ src/bin/train/spectrogram_model/trainer.py
@@ -171,8 +171,9 @@ class Trainer():
         # NOTE: GCP shutdowns do not trigger `atexit`; therefore, it's useful to always save
         # a temporary checkpoint just in case.
         checkpoint_path = self.save_checkpoint()
-        if hasattr(self,
-                   '_last_repeat_timer_checkpoint') and self._last_repeat_timer_checkpoint.exists():
+        if (hasattr(self, '_last_repeat_timer_checkpoint') and
+                self._last_repeat_timer_checkpoint is not None and
+                self._last_repeat_timer_checkpoint.exists()):
             logger.info('Unlinking temporary checkpoint: %s',
                         str(self._last_repeat_timer_checkpoint))
             self._last_repeat_timer_checkpoint.unlink()
@@ -225,10 +226,11 @@ class Trainer():
         """ Save a checkpoint.
 
         Returns:
-            (pathlib.Path): Path the checkpoint was saved to.
+            (pathlib.Path or None): Path the checkpoint was saved to or None if checkpoint wasn't
+                saved.
         """
         if src.distributed.is_master():
-            return Checkpoint(
+            checkpoint = Checkpoint(
                 comet_ml_project_name=self.comet_ml.project_name,
                 directory=self.checkpoints_directory,
                 model=(self.model.module if src.distributed.is_initialized() else self.model),
@@ -236,7 +238,10 @@ class Trainer():
                 input_encoder=self.input_encoder,
                 epoch=self.epoch,
                 step=self.step,
-                comet_ml_experiment_key=self.comet_ml.get_key()).save()
+                comet_ml_experiment_key=self.comet_ml.get_key())
+            if checkpoint.path.exists():
+                return None
+            return checkpoint.save()
         else:
             return None
